06-workshops/01-AgentCore-runtime/05-hosting-a2a/02-a2a-agent-sigv4/requirements.txt
boto3==1.42.28
bedrock-agentcore==1.2.0
bedrock-agentcore-starter-toolkit
strands-agents[a2a]
strands-agents-tools
httpx
